% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MethylSeekRGR-class.R
\name{as}
\alias{MethylSeekRGR}
\alias{as}
\title{Coerce a MethPat object of CG 1-tuples to a list of MethylSeekR-compatible
GRanges object(s), one list element per sample.}
\arguments{
\item{methpat}{A \code{\link[MethylationTuples]{MethPat}} object containing CpG
1-tuples. The \code{\link[MethylationTuples]{MethPat}} object should already
been processed with
\code{MethylationTuples::\link[MethylationTuples]{filterOutVariants}} and
\code{MethylationTuples::\link[MethylationTuples]{collapseStrand}}.}
}
\value{
A list of \code{\link{MethylSeekRGR}} objects, which are
compatible with the \code{MethylSeekR} Bioconductor package.
}
\description{
Coerce a MethPat object of CG 1-tuples to a list of MethylSeekR-compatible
GRanges object(s), one list element per sample.
}

