% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findMostOverlapping.R
\name{findMostOverlapping}
\alias{findMostOverlapping}
\title{Find "most overlapping" range in subject for each element of query.}
\usage{
findMostOverlapping(query, subject)
}
\arguments{
\item{query}{Any object for which a `findOverlaps` method is defined, e.g.,
a \code{\link[GenomicRange]{GRanges}}.}

\item{subject}{Any object for which a `findOverlaps` method is defined, e.g.,
a \code{\link[GenomicRange]{GRanges}}.}
}
\value{
An \code{integer} vector containing for each \code{query} the index
of the \code{subject} that it most overlaps.
}
\description{
Find "most overlapping" range in subject for each element of query.
}
\details{
Ties are broken at random.
}

