% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funByPM.R
\name{funByPM}
\alias{funByPM}
\title{Apply a function to a
\code{MethylationTuples::\link[MethylationTuples]{MethPat}} object,
with results stratified by the region type of the accompanying
\code{\link{PartitionedMethylome}}.}
\usage{
funByPM(FUN, pm, methpat, min_cov, ...)
}
\arguments{
\item{FUN}{The function to be applied to each element of \code{pm}: see
'Details'. Only a limited number of functions are currently supported,
specifically \code{MethylationTuples::\link[MethylationTuples]{methLevel}()},
\code{MethylationTuples::\link[MethylationTuples]{cometh}()}, and
\code{MethylationTuples::\link[MethylationTuples]{patternFreqs}()}.}

\item{pm}{A \code{\link{PartitionedMethylome}} object.}

\item{methpat}{A \code{MethylationTuples::\link[MethylationTuples]{MethPat}}
object.}

\item{min_cov}{An \code{integer} specifying the minimum coverage required
in order to use an m-tuple in the analysis.}

\item{...}{Optional arguments to \code{FUN}.}
}
\value{
A \code{data.table::\link[data.table]{data.table}}.
}
\description{
Apply a function to a
\code{MethylationTuples::\link[MethylationTuples]{MethPat}} object,
with results stratified by the region type of the accompanying
\code{\link{PartitionedMethylome}}.
}

