% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateMethylome2-class.R
\docType{methods}
\name{methinfo,SimulatedMethylome2-method}
\alias{methinfo}
\alias{methinfo,SimulatedMethylome2-method}
\title{Extract MethInfo}
\usage{
\S4method{methinfo}{SimulatedMethylome2}(object)
}
\description{
Extract MethInfo
}

