% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulatedMethylome-class.R
\docType{methods}
\name{methLevel,SimulatedMethylome-method}
\alias{methLevel}
\alias{methLevel,SimulatedMethylome-method}
\title{Compute methylation levels.}
\usage{
\S4method{methLevel}{SimulatedMethylome}(object, statistic = c("beta-values",
  "M-values"), offset = 1L)
}
\arguments{
\item{object}{A \code{\link{SimulatedMethylome}} object.}

\item{statistic}{A \code{character} string indicating which methylation
level statistic is to be computed. One of "\code{beta-values}" or
"\code{M-values}" (see below).}

\item{offset}{A \code{numeric} vector with length 1 used when computing
M-values (default: 1).}
}
\value{
A \code{\link[base]{matrix}}, with the same dimensions and dimension
names as \code{x}, of methylation levels at each methylation loci in each
sample.
}
\description{
Compute methylation levels.
}
\details{
TODO: Define beta-values and M-values. Note any differences with how others
define beta-values or M-values, e.g., minfi.
}

