% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methsim.R
\docType{package}
\name{methsim}
\alias{methsim}
\alias{methsim-package}
\title{Simulate DNA methylation data.}
\description{
\pkg{methsim} is software to simulate DNA methylation sequencing data.
\code{methsim} can currently simulate data from bisulfite-sequencing assays,
such as methylC-seq or BS-seq. It uses a non-stationary, inhomogeneous Markov
model. This model can incorporate the strong spatial dependence of DNA
methylation. Simulation parameters can be estimated from data or specified
by the user.
}
\details{
Please refer to the vignettes to see how to use the \pkg{methsim} package.
}

