% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MethSimParam-class.R
\docType{class}
\name{MethSimParam-class}
\alias{MethSimParam}
\alias{MethSimParam-class}
\title{MethSimParam class}
\description{
An S4 class for the parameters used by \code{\link{simulate}}.
}

