% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MethylomeParam-class.R
\docType{class}
\name{MethylomeParam-class}
\alias{MethylomeParam}
\alias{MethylomeParam-class}
\title{MethylomeParam class}
\description{
An S4 class for the parameters used by
\code{\link{simulate,MethylomeParam-method}}.
}

