% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PartitionedMethylome-class.R
\docType{class}
\name{PartitionedMethylome-class}
\alias{PartitionedMethylome}
\alias{PartitionedMethylome-class}
\title{PartitionedMethylome class.}
\description{
PartitionedMethylome class.
}

