% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partitionMethylome}
\alias{partitionMethylome}
\title{Partition a methylome into UMRs, LMRs, PMRs and others.}
\usage{
partitionMethylome(umrs_lmrs, pmrs)
}
\arguments{
\item{umrs_lmrs}{The output of
\code{MethylSeekR::\link[MethylSeekR]{segmentUMRsLMRs}}.}

\item{pmrs}{The output of \code{MethylSeekR\link[MethylSeekR]{segmentPMDs}}.}

\item{bsgenome}{A \code{BSgenome::\link[BSgenome]{BSgenome}} object of the
relevant organism.}
}
\value{
A \code{\link{PartitionedMethylome}}.
}
\description{
Paritition a methylome into unmethylated regions (UMRs), lowly-methylated
regions (LMRs), partially-methylated regions (PMRs) and mostly-methylated
regions (others) based on the beta-values.
}

