% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MethylomeParam-class.R
\docType{methods}
\name{simulate,MethylomeParam-method}
\alias{simulate,MethylomeParam-method}
\title{Simulate a methylome.}
\usage{
\S4method{simulate}{MethylomeParam}(object, nsim = 1, seed = NULL,
  epsilon = 0.01, seqlevels, BPPARAM = bpparam(), ...)
}
\arguments{
\item{object}{A \code{\link{MethylomeParam}} object.}

\item{nsim}{The number of methylomes to simulate using the
parameters given in \code{object}.}

\item{seed}{An object specifying if and how the random number generator
should be initialized ('seeded'). For the \code{MethylomeParam}
method, either \code{NULL} or an integer that will be used in a call to
\code{base::\link[base]{set.seed}} before simulating the samples
(methylomes). If set, the value is saved as the "\code{seed}" attribute of
the returned value. The default, \code{NULL}, will not change the random
generator state, and return \code{\link{.Random.seed}} as the "\code{seed}"
attribute, see 'Value'.}

\item{epsilon}{An offset added/subtracted to a region's sampled methylation
level to avoid zero/one values.}

\item{seqlevels}{A character vector of
\code{GenomeInfoDb::\link[GenomeInfoDb]{seqlevels}} at which to simulate a
methylome. If missing, the default is to use all available seqlevels.}

\item{BPPARAM}{An optional
\code{BiocParallel::\link[BiocParallel]{BiocParallelParam}} instance
determining the parallel back-end to be used during evaluation.}

\item{...}{Additional arguments passed to the
\code{methsim:::.sampleComethDT()}.}
}
\value{
A list of length \code{nsim} of SimulatedMethylome objects.
}
\description{
\code{simulate()} samples the distribution of average methylation level
(\code{MethLevelDT} slot) and the distributions of within-fragment
co-methylation (\code{ComethDT} slot) to compute the transition
probabilities under a first-order Markov process.
}
\note{
Currently only support simulation of CpG methylation and unstranded
methylomes.
}
\section{Co-methylation}{

\strong{TODO}: Describe  \code{mean_fun} and\code{sd_fun}.
}

