% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateBSParam-class.R
\docType{methods}
\name{simulate,SimulateBSParam-method}
\alias{simulate,SimulateBSParam-method}
\title{Simulate a bisulfite-sequencing experiment.}
\usage{
\S4method{simulate}{SimulateBSParam}(object, nsim = 1, seed = NULL,
  BPPARAM = bpparam(), sequencingType = "SE", readLength = 100L, ...)
}
\arguments{
\item{object}{a \code{\link{SimulateBSParam}} object.}

\item{nsim}{the number of samples to simulate using the parameters given in
\code{object}.}

\item{seed}{an object specifying if and how the random number generator
should be initialized ('seeded'). For the "MethSimParam" method, either
\code{NULL} or an integer that will be used in a call to
\code{base::\link[base]{set.seed}} before simulating the samples. If set,
the value is saved as the "\code{seed}" attribute of the returned value. The
default, \code{NULL}, will not change the random generator state, and return
\code{\link{.Random.seed}} as the "\code{seed}" attribute, see 'Value'.}

\item{BPPARAM}{an optional
\code{BiocParallel::\link[BiocParallel]{BiocParallelParam}} instance
determining the parallel back-end to be used during evaluation.}

\item{sequencingType}{a character specifying either single-end ('SE') or
paired-end ('PE'). \strong{NOTE}: Only single-end data is currently
supported.}

\item{readLength}{an integer specifying the read length.}
}
\value{
a SimulatedBS object.
}
\description{
Simulate a bisulfite-sequencing experiment.
}
\note{
Currently only simulates whole-genome bisulfite-sequencing data.
\strong{WARNING}: Currently reads are simulated for circular seqlevels
such as 'chrM' (mitochondrial DNA).
}

