% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateMethylomeParam-class.R
\docType{methods}
\name{simulate,SimulateMethylomeParam-method}
\alias{simulate,SimulateMethylomeParam-method}
\title{Simulate a methylome.}
\usage{
\S4method{simulate}{SimulateMethylomeParam}(object, nsim = 1, seed = NULL,
  comethylation_function = .sampleComethDTLoess, BPPARAM = bpparam(), ...)
}
\arguments{
\item{object}{A \code{\link{SimulateMethylomeParam}} object.}

\item{nsim}{Number of samples to simulate using the parameters given in
\code{object}.}

\item{seed}{An object specifying if and how the random number generator
should be initialized ('seeded'). For the "MethSimParam" method, either
\code{NULL} or an integer that will be used in a call to
\code{base::\link[base]{set.seed}} before simulating the samples. If set,
the value is saved as the "\code{seed}" attribute of the returned value. The
default, \code{NULL}, will not change the random generator state, and return
\code{\link{.Random.seed}} as the "\code{seed}" attribute, see 'Value'.}

\item{BPPARAM}{An optional
\code{BiocParallel::\link[BiocParallel]{BiocParallelParam}} instance
determining the parallel back-end to be used during evaluation.}

\item{...}{additional arguments passed to the \code{comethylation_function}.}

\item{comethylation_model}{the type}
}
\value{
A SimulatedMethylome, which is the underlying "true" methylome for
a single sample.
}
\description{
Simulate a methylome.
}
\note{
Currently only simulates CpG methylation.
}

