% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{simulate}
\alias{simulate}
\title{Simulate responses}
\usage{
simulate(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{an object representing the simulation parameters.}

\item{nsim}{number of responses to simulate. Defaults to 1.}

\item{seed}{an object specifying if and how the random number generator
should be initialized ('seeded'). Either \code{NULL} or an integer that will
be used in a call to \code{base::\link[base]{set.seed}} before simulating
the responses. If set, the value is saved as the "\code{seed}" attribute of
the returned value. The default, \code{NULL} will not change the random
generator state, and return \code{base::\link[base]{.Random.seed}} as the
"\code{seed}" attribute, see 'Value'.}

\item{...}{additional optional arguments.}
}
\value{
See \code{?stats::\link[stats]{simulate}} for the value returned by
the default method.

Specific methods defined in other packages should behave as consistently as
possible with the default method.
}
\description{
Simulate data according to given parameters.

NOTE: This man page is for the \code{simulate} \emph{S4 generic function}
defined in the \pkg{methsim} package. See
\code{?statistics::\link[stats]{simulate}} for the default method
(defined in the \pkg{statistics} package). Other packages can define
specific S4 methods for objects not supported by the default method.
}
\examples{
\dontrun{
# TODO
}
}
\seealso{
\itemize{
 \item \code{stats::\link[stats]{simulate}} for the default \code{simulate}
   method.
 \item \code{\link[methods]{showMethods}} for displaying a summary of the
   methods defined for a given generic function.
 \item \code{\link[methods]{selectMethod}} for getting the definition of
   a specific method.
 \item \code{\link{simulate,SimulateMethylomeParam-method}} for an example
 of a specific \code{simulate} method (defined for
 \code{\link{SimulateMethylomeParam}} objects).
 \item \link[BiocGenerics]{BiocGenerics} for a summary of all the generics
 defined in the \pkg{BiocGenerics} package.
}
}
\keyword{methods}

