% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateBSParam2-class.R
\docType{methods}
\name{simulate2,SimulateBSParam2-method}
\alias{simulate2,SimulateBSParam2-method}
\title{Simulate a bisulfite-sequencing experiment using the second method.}
\usage{
\S4method{simulate2}{SimulateBSParam2}(object, nsim = 1, seed = NULL,
  BPPARAM = bpparam(), sequencingType = "SE", readLength = 100L,
  seqlevels, simplify = 0L, ...)
}
\arguments{
\item{object}{a \code{\link{SimulateBSParam2}} object.}

\item{nsim}{the number of samples to simulate using the parameters given in
\code{object}.}

\item{seed}{an object specifying if and how the random number generator
should be initialized ('seeded'). For the "MethSimParam" method, either
\code{NULL} or an integer that will be used in a call to
\code{base::\link[base]{set.seed}} before simulating the samples. If set,
the value is saved as the "\code{seed}" attribute of the returned value. The
default, \code{NULL}, will not change the random generator state, and return
\code{\link{.Random.seed}} as the "\code{seed}" attribute, see 'Value'.}

\item{BPPARAM}{an optional
\code{BiocParallel::\link[BiocParallel]{BiocParallelParam}} instance
determining the parallel back-end to be used during evaluation.}

\item{sequencingType}{a character specifying either single-end ('SE') or
paired-end ('PE'). \strong{NOTE}: Only single-end data is currently
supported.}

\item{readLength}{an integer specifying the read length.}

\item{seqlevels}{A vector of seqlevels at which to bisulfite-sequencing
reads. Will use all available seqlevels if missing, where availability is
defined by the seqlevels "in use" in the SimulatedMethylome slot.}

\item{simplify}{An integer specifying whether, and by how much, the object
should be simplified. The default value, \code{simplify = 0}, means no
simplification and the returned object is a \code{\link{SimulatedBS}} object.
A value greater than zero simplifies the returned object as a
\code{MethylationTuples::\link[MethylationTuples]{MethPat}} object
containing tuples with \code{size = simplify}, e.g., \code{simplify = 1}
will return a \code{MethylationTuples::\link[MethylationTuples]{MethPat}}
containing 1-tuples.}
}
\value{
a SimulatedBS2 object.
}
\description{
Simulate a bisulfite-sequencing experiment using the second method.
}
\note{
Currently only simulates whole-genome bisulfite-sequencing data.
\strong{WARNING}: Currently reads are simulated for circular seqlevels
such as 'chrM' (mitochondrial DNA).
}

