% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateMethylomeParam-class.R
\docType{methods}
\name{simulate2,SimulateMethylomeParam-method}
\alias{simulate2,SimulateMethylomeParam-method}
\title{Simulate a methylome the second method.}
\usage{
\S4method{simulate2}{SimulateMethylomeParam}(object, nsim = 1, seed = NULL,
  epsilon = 0.01, comethylation_function, seqlevels, mc_order = 1L, ...)
}
\arguments{
\item{object}{A \code{\link{SimulateMethylomeParam}} object.}

\item{nsim}{The number of samples to simulate using the parameters given in
\code{object}.}

\item{seed}{An object specifying if and how the random number generator
should be initialized ('seeded'). For the "MethSimParam" method, either
\code{NULL} Or an integer that will be used in a call to
\code{base::\link[base]{set.seed}} before simulating the samples. If set,
the value is saved as the "\code{seed}" attribute of the returned value. The
default, \code{NULL}, will not change the random generator state, and return
\code{\link{.Random.seed}} as the "\code{seed}" attribute, see 'Value'.}

\item{epsilon}{An offset added/subtracted to a region's sampled methylation
level should it be one/zero values.}

\item{comethylation_function}{A function used to sample from the
co-methylation distribution.}

\item{seqlevels}{A vector of seqlevels for which to simulate a methylome. The
default uses all available seqlevels.}

\item{mc_order}{The order of the Markov chain. Currently only a value of
1 is supported.}

\item{...}{Additional arguments passed to the \code{comethylation_function}.}
}
\value{
A SimulatedMethylome, which is the underlying "true" methylome for
a single sample.
}
\description{
Rather than using weights (PatternFreqsDT) and simulating the
'true' methylome directly (as a SimulatedMethylome object), simulate2()
samples the distribution of average methylation levels (MethLevelDT)
and the distributions of within-fragment co-methylation (ComethDT)
and computes the resulting transition probabilities under a first-order
Markov process.
}
\note{
Currently only simulates CpG methylation.

Currently only support simulation of CpG methylation and unstranded
methylomes.

TODO: Random number generation is not yet properly implemented and so
results will likely not be reproducible even if the same seed is used.
}

