% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateBSParam-class.R
\docType{class}
\name{SimulateBSParam-class}
\alias{SimulateBSParam}
\alias{SimulateBSParam-class}
\title{SimulateBSParam class}
\description{
An S4 class for the parameters used by
\code{\link{simulate,SimulateBSParam-method}}.
}

