% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateBSParam2-class.R
\docType{class}
\name{SimulateBSParam2-class}
\alias{SimulateBSParam2}
\alias{SimulateBSParam2-class}
\title{SimulateBSParam2 class}
\description{
An S4 class for the parameters used by
\code{\link{simulate2,SimulateBS2Param-method}}.
}

