% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulatedBS-class.R
\docType{class}
\name{SimulatedBS-class}
\alias{SimulatedBS}
\alias{SimulatedBS-class}
\title{SimulatedBS class.}
\description{
SimulatedBS class.
}
\details{
An S4 class to store a single simulated bisulfite-sequencing
experiment.
}

