% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulatedMethylome2-class.R
\docType{class}
\name{SimulatedMethylome2-class}
\alias{SimulatedMethylome2}
\alias{SimulatedMethylome2-class}
\title{SimulatedMethylome2 class.}
\description{
SimulatedMethylome2 class.
}
\details{
An S4 class to store a simulated methylomes.
}

