% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateMethylomeParam-class.R
\docType{class}
\name{SimulateMethylomeParam-class}
\alias{SimulateMethylomeParam}
\alias{SimulateMethylomeParam-class}
\title{SimulateMethylomeParam class}
\description{
An S4 class for the parameters used by
\code{\link{simulate,SimulateMethylome-method}}.
}

