% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WGBSParam-class.R
\docType{class}
\name{WGBSParam-class}
\alias{WGBSParam}
\alias{WGBSParam-class}
\title{WGBSParam class}
\description{
An S4 class for the parameters used by
\code{\link{simulate,WGBSParam-method}}.
}

