% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as}
\alias{MethylSeekRGR}
\alias{as}
\title{Coerce a MethPat object of CG 1-tuples to MethylSeekR-compatible GRanges
object(s).}
\arguments{
\item{methpat}{A \code{\link[MethylationTuples]{MethPat}} object containing CpG
1-tuples. The \code{\link[MethylationTuples]{MethPat}} object should already
been processed with
\code{MethylationTuples::\link[MethylationTuples]{filterOutVariants}} and
\code{MethylationTuples::\link[MethylationTuples]{collapseStrand}}.}
}
\value{
A list of \code{\link{MethylSeekRGR}} objects, which are
compatible with the \code{MethylSeekR} Bioconductor package.
}
\description{
Coerce a MethPat object of CG 1-tuples to MethylSeekR-compatible GRanges
object(s).
}

