% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findMostOverlapping}
\alias{findMostOverlapping}
\title{Find "most overlapping" range in subject for each element of query}
\usage{
findMostOverlapping(query, subject)
}
\arguments{
\item{query}{Any object for which a `findOverlaps` method is defined, e.g.,
a \code{\link[GenomicRange]{GRanges}}.}

\item{subject}{Any object for which a `findOverlaps` method is defined, e.g.,
a \code{\link[GenomicRange]{GRanges}}.}
}
\value{
An \code{integer} vector.
}
\description{
Find "most overlapping" range in subject for each element of query
}

