% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MethylSeekRGR-class.R
\docType{class}
\name{MethylSeekRGR-class}
\alias{MethylSeekRGR}
\alias{MethylSeekRGR-class}
\title{MethylSeekRGR class}
\description{
An S4 class to formalise the informal
\code{GenomicRanges::\link[GenomicRanges]{GRanges}}-based class used by
\code{\link[MethylSeekR]{MethylSeekR-package}}.
}

