% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partitionMethylome}
\alias{partitionMethylome}
\title{Partition a methylome into UMRs, LMRs, PMRs and MMRs.}
\usage{
partitionMethylome(umrs_lmrs, pmrs)
}
\arguments{
\item{umrs_lmrs}{The output of
\code{MethylSeekR\link[MethylSeekR]{segmentUMRsLMRs}}.}

\item{pmrs}{The output of \code{MethylSeekR\link[MethylSeekR]{segmentPMDs}}.}
}
\value{
A \code{\link{PartitionedMethylome}}.
}
\description{
Paritition a methylome into unmethylated regions (UMRs), lowly-methylated
regions (LMRs), partially-methylated regions (PMRs) and mostly-methylated
regions (MMRs) based on the beta-values.
}
\details{
This function is a convenience function wrapping around several functions
provided by the \code{MethylSeekR} Bioconductor package. It is not a
replacement.
}

