% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partitionMethylome.R
\name{partitionMethylome}
\alias{partitionMethylome}
\title{Partition a methylome into UMRs, LMRs, PMRs and others.}
\usage{
partitionMethylome(umrs_lmrs, pmrs)
}
\arguments{
\item{umrs_lmrs}{The output of
\code{MethylSeekR::\link[MethylSeekR]{segmentUMRsLMRs}}.}

\item{pmrs}{The output of \code{MethylSeekR\link[MethylSeekR]{segmentPMDs}}.}
}
\value{
A \code{\link{PartitionedMethylome}}.
}
\description{
Paritition a methylome into unmethylated regions (UMRs), lowly-methylated
regions (LMRs), partially-methylated regions (PMRs) and mostly-methylated
regions (others) based on the beta-values.
}

