% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BSParam-class.R
\docType{methods}
\name{simulate,BSParam-method}
\alias{simulate,BSParam-method}
\title{Simulate a bisulfite-sequencing experiment.}
\usage{
\S4method{simulate}{BSParam}(object, nsim = 1, seed = NULL, seqlevels,
  simplify = 0L, BPPARAM = bpparam())
}
\arguments{
\item{object}{A \code{\link{BSParam}} object.}

\item{nsim}{The number of samples to simulate using the parameters given in
\code{object}. Additional samples will be technical replicates.}

\item{seed}{An object specifying if and how the random number generator
should be initialized ('seeded'). For the "BSParam" method, either
\code{NULL} or an integer that will be used in a call to
\code{base::\link[base]{set.seed}} before simulating the samples. If set,
the value is saved as the "\code{seed}" attribute of the returned value. The
default, \code{NULL}, will not change the random generator state, and return
\code{\link{.Random.seed}} as the "\code{seed}" attribute, see 'Value'.}

\item{seqlevels}{A character vector of
\code{GenomeInfoDb::\link[GenomeInfoDb]{seqlevels}} at which to
bisulfite-sequencing reads. If missing, the default is to use all available
seqlevels.}

\item{simplify}{An integer specifying whether, and by how much, the object
should be simplified, see 'Value'.}

\item{BPPARAM}{an optional
\code{BiocParallel::\link[BiocParallel]{BiocParallelParam}} instance
determining the parallel back-end to be used during evaluation.}
}
\value{
The return value depends on the value of \code{simplify}. The
default (\code{simplify = 0}) means no simplification and the returned
object is a \code{list} of length \code{nsim} containing
\code{\link{SimulatedBS}} objects. If \code{simplify} is greater than zero,
the returned object is a
\code{MethylationTuples::\link[MethylationTuples]{MethPat}} object with
\code{\link[MethylationTuples]{size} = simplify} and \code{ncol = simplify}.
}
\description{
Simulate a bisulfite-sequencing experiment.
}
\note{
Currently only simulates whole-genome bisulfite-sequencing data.
}
\section{Warnings}{

\itemize{
 \item Reads are \strong{not} yet simulated for circular seqlevels such as
       'chrM' (mitochondrial DNA).
 \item Only single-end sequencing ('SE') is currently supported.
 \item This is currently not RNG-safe since random numbers are generated
 within the parallel process and at the \code{Rcpp} level. Therefore,
 results may not be reproducible, even given the same \code{seed}.
 \strong{This is a work in progress and will be fixed.}
 \item The \code{nsim} simulations are currently simulated in
 \strong{serial} (but steps of each simulation may be run in parallel).
}
}

