% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SimulateMethylome2-class.R
\docType{class}
\name{SimulatedMethylome2-class}
\alias{SimulatedMethylome}
\alias{SimulatedMethylome2-class}
\title{SimulatedMethylome class.}
\description{
SimulatedMethylome class.
}
\details{
An S4 class to store a single simulated methylome.
}

